feat: integrate shell stub generation with search command and remove duplicate functionality
- Add --shell-stubs flag to search command for filtered shell function generation
- Move create_shell_function_name utility to shared helper module with tests
- Remove obsolete create-shell-stubs command to eliminate code duplication

The new approach allows commands like:
- bkmr search --tags="database" --shell-stubs
- bkmr search "backup" --shell-stubs
- bkmr search --limit=5 --shell-stubs
