+++ bkmr/src/cli/bookmark_commands.rs
@@ -1218,102 +1218,4 @@ mod tests {
     }
 }
 
-/// Create shell function stubs for all shell script bookmarks
-#[instrument(skip(cli))]
-pub fn create_shell_stubs(cli: Cli) -> CliResult<()> {
-    if let Commands::CreateShellStubs = cli.command.unwrap() {
-        let bookmark_service = create_bookmark_service();
-        
-        // Get all bookmarks
-        let bookmarks = bookmark_service.get_all_bookmarks(None, None)?;
-        
-        // Filter for shell script bookmarks
-        let shell_bookmarks: Vec<_> = bookmarks
-            .into_iter()
-            .filter(|bookmark| {
-                bookmark.tags.iter().any(|tag| tag.is_system_tag_of(SystemTag::Shell))
-            })
-            .collect();
-        
-        // Generate shell function stubs
-        for bookmark in shell_bookmarks {
-            if let Some(id) = bookmark.id {
-                // Create a valid shell function name from the bookmark title
-                let function_name = create_shell_function_name(&bookmark.title);
-                
-                // Output the shell function
-                println!("{}() {{ bkmr open --no-edit {} -- \"$@\"; }}", function_name, id);
-                println!("export -f {}", function_name);
-            }
-        }
-    }
-    Ok(())
-}
 
-/// Create a valid shell function name from a bookmark title
-fn create_shell_function_name(title: &str) -> String {
-    let cleaned_name = title
-        .chars()
-        .map(|c| {
-            if c.is_alphanumeric() || c == '-' {
-                c.to_ascii_lowercase()  // Preserve hyphens
-            } else if c.is_whitespace() || c == '_' {
-                '_'  // Only spaces and underscores become underscores
-            } else {
-                // Skip other invalid characters
-                '\0'
-            }
-        })
-        .filter(|&c| c != '\0')
-        .collect::<String>()
-        .trim_matches('_')
-        .to_string();
-    
-    // Ensure we have a valid function name
-    if cleaned_name.is_empty() {
-        "shell_script".to_string()
-    } else if cleaned_name.chars().next().unwrap().is_ascii_digit() {
-        // Shell function names can't start with a digit
-        format!("script-{}", cleaned_name)
-    } else {
-        cleaned_name
-    }
-}
-
-#[cfg(test)]
-mod shell_stubs_tests {
-    use super::*;
-
-    #[test]
-    fn test_create_shell_function_name() {
-        // Normal case
-        assert_eq!(create_shell_function_name("Deploy Script"), "deploy_script");
-        
-        // Hyphens should be preserved
-        assert_eq!(create_shell_function_name("backup-database"), "backup-database");
-        
-        // Mixed hyphens and spaces
-        assert_eq!(create_shell_function_name("My-Awesome Script!"), "my-awesome_script");
-        
-        // Starting with number
-        assert_eq!(create_shell_function_name("2fa Setup"), "script-2fa_setup");
-        
-        // Starting with number and hyphens
-        assert_eq!(create_shell_function_name("2fa-backup"), "script-2fa-backup");
-        
-        // Empty after cleaning
-        assert_eq!(create_shell_function_name("!@#$%"), "shell_script");
-        
-        // Already valid
-        assert_eq!(create_shell_function_name("backup"), "backup");
-        
-        // Already valid with hyphen
-        assert_eq!(create_shell_function_name("backup-db"), "backup-db");
-        
-        // Mixed case with underscores
-        assert_eq!(create_shell_function_name("Database_Backup_Script"), "database_backup_script");
-        
-        // Mixed hyphens, underscores, and spaces
-        assert_eq!(create_shell_function_name("Deploy-DB Script_v2"), "deploy-db_script_v2");
-    }
-}
